backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dj-rest-auth[with-social]>=7.0.1",
    "django-cors-headers>=4.7.0",
    "django>=5.1.6",
    "djangorestframework>=3.15.2",
    "django-allauth[socialaccount]>=65.6.0",
    "dotenv>=0.9.9",
]
